requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
urllib3==2.1.0
wheel==0.42.0